+++ TetrisAPI/IInputManager.h
@@ -8,6 +8,6 @@ namespace TetrisAPI
 	{
 	public:
 		virtual bool Check(EInputType inputType) const = 0;
-		virtual void Register(EInputType inputType, int key) = 0;
+		virtual void Register(EInputType inputType, uint16_t key) = 0;
 	};
 }
